binding using udf
